tests/wpt/tests/html/semantics/interactive-elements/the-dialog-element/dialog-focusability.html

<!doctype html>
<meta charset="utf-8">
<title>dialog element: focusability</title>
<link rel=help href="https://github.com/whatwg/html/pull/8199">
<link rel=author href="mailto:masonf@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<button id="before">before</button>
<dialog id="dialog" open>
  <button id="within">button</button>
</dialog>
<button id="after">after</button>

<script>
  function navigateForward() {
    const TAB = '\ue004';
    return test_driver.send_keys(document.body, TAB);
  }
  async function assert_focus_order(elements) {
    assert_true(elements.length >= 2);
    elements[0].focus();
    for(let i=0;i<elements.length;++i) {
      assert_equals(document.activeElement,elements[i],`Focus order mismatch at step ${i+1}/${elements.length}`);
      await navigateForward();
    }
  }

  async_test((t) => {
    window.onload = async () => {
      await assert_focus_order([before,within,after]);
      t.done();
    };
  }, "The dialog element itself should not be keyboard focusable.");
</script>
